Fix: Move data fetching from AuthorsListViewModel to MainViewModel
Fixes bug where authors were being downloaded every time AuthorsListScreen
opened. Data fetching now happens once in MainViewModel init block instead
of on every screen resume.

Changes:
- MainViewModel: Added authorRepository and worksRepository injection,
  fetch data in init block
- AuthorsListViewModel: Removed onViewResumed() override, hasInitiallyFetched
  flag, and initial fetch methods. Set isLoading=false in observeAuthors()
